% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{file_merge}
\alias{file_merge}
\title{file merge}
\usage{
file_merge(seaurat, file_site)
}
\arguments{
\item{seaurat}{object}

\item{file_site}{folder}
}
\value{
seurat object
}
\description{
file merge
}
\examples{
file_merge(data,file_site)
}
