% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{load_data}
\alias{load_data}
\title{load data}
\usage{
load_data(file_site, sample = T, data_type = "rds")
}
\arguments{
\item{file_site}{File name or folder}

\item{sample}{T or F}

\item{data_type}{single cell data type,such as:rds h5 loom 10X}
}
\value{
seurat object
}
\description{
load data
}
\examples{
load_data(file_site,sample=T,data_type='rds')
}
